Setup OpenSaml in custom bean with custom properties
